split functions from memory TU
